Added optional Fourier transform support to the gettile command
As explained in the README, the FFT support is implemented with the
NFFTv3 library, and opting in to FFT support is as simple as adding
FFT_SUPPORT=1 as an argument to the make invocation when building the
project.  If such support is enabled, then calling gettile with the
channel name replaced by name.DFT will print a JSON string containing
the a dft field, which holds the amplitude of the discrete Fourier
transform as calculated by NFFTv3, scaled and packed as UTF-8 characters.
Currently we scale into 256 bins, so all values smaller than half the max
turn into characters with value strictly less than 128, and are printed
as single-byte characters.  All other values require multiple bytes.