Removed NFFT, switched to NFFT with interpolated data, result now 2D
Now we interpolate data to the 10 Hz level, then calculate the FFT of
that data in overlapping windows.  We make the window the width of a
tile, centered on the region that we will draw as that window, and
compute four windows per tile.

This is a demo-ready version, but it has some shortcomings:

1) It uses the tile that would be returned for the normal tile fetch in
order to calculate the DFT, rather than using the raw data.  This means
that averaged data is used instead of the original data, and the
high-frequency variation is lost.  Also, the windows that should extend
beyond the end of the tile do not actually get the data they need.
2) There are only four windows per tile.  The standard tile fetcher of
the grapher tries to pull tiles that are about the width of the
displayed region, which leads to an overly coarse display in the grapher.
3) It calculates the DFT on the fly, rather than precomputing and
storing the DFT data.  This makes tile fetch slower.