Switched to precomputation of DFT tiles
Now there is a second channel to hold the DFT computed from an original
channel, and a new genfft executable that actually precomputes that DFT.
The gettile executable thus no longer needs FFT support directly, and
only reads the tile from disk.

The computed DFT, after it is adjusted for presentation to the outside
world, is saved to disk in a flattened format.  In other words, the
two-dimensional array of data points in the DFT tile is flattened into a
one-dimensional array of double-valued data points, and one
string-valued data point is saved to store enough information to
reconstitute those points.  The MapSerializer class is used to serialize
and deserialize a std::map<std::string, std::string> into a string
format, keeping that information intact between an execution of the
genfft app and an execution of the gettile app.

The genfft app and gettile app are not perfect - the genfft app should be
cognizant of the root tile containing all data points, just as the
Channel class is, and the gettile app should take lower-level DFT tiles
if no high-level DFT tiles are available - but this implementation is
nearly complete.